refactor: 설정 부분 리팩토링
