+++ src/main/java/until/the/eternity/dcs/common/notification/RedisSender.java
@@ -1,7 +1,5 @@
-package until.the.eternity.dcs.domain.notice.infrastructure;
+package until.the.eternity.dcs.common.notification;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.HashMap;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -10,14 +8,12 @@ import org.springframework.data.redis.connection.stream.RecordId;
 import org.springframework.data.redis.connection.stream.StreamRecords;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
-import until.the.eternity.dcs.domain.notice.dto.stream.NotificationJob;
+import until.the.eternity.dcs.common.notification.dto.NotificationJob;
 
 @Service
 @RequiredArgsConstructor
 public class RedisSender {
-
     private final StringRedisTemplate redisTemplate;
-    private final ObjectMapper om = new ObjectMapper();
 
     @Value("${app.notification.stream}")
     private String stream;
@@ -26,16 +22,12 @@ public class RedisSender {
     private long maxLen;
 
     public RecordId enqueue(NotificationJob job) {
-        Map<String, String> f = new HashMap<>();
-        f.put("notificationId", job.notificationId());
-        f.put("userId", String.valueOf(job.userId()));
-        f.put("channel", job.channel());
-        f.put("noticeType", job.noticeType().getCode());
-        f.put("url", job.url());
+        Map<String, String> record = job.toMap();
 
         XAddOptions opts = XAddOptions.maxlen(maxLen).approximateTrimming(true);
+
         return redisTemplate
                 .opsForStream()
-                .add(StreamRecords.mapBacked(f).withStreamKey(stream), opts);
+                .add(StreamRecords.mapBacked(record).withStreamKey(stream), opts);
     }
 }
